requirements.txt
aiogram==3.1.1
aiogram_dialog==2.0.0
python-dotenv==1.0.0
